perf(webapp) stackedBarChart refactor code and componetization
